more double/Money conversion; using destination currency's decimal places in the numeric input

+++ src/com/money/manager/ex/core/Core.java
@@ -280,9 +280,9 @@ public class Core {
         CurrencyService currencyService = new CurrencyService(mContext);
 
         if (currencyId == null) {
-            view.setText(currencyService.getBaseCurrencyFormatted(amount.toDouble()));
+            view.setText(currencyService.getBaseCurrencyFormatted(amount));
         } else {
-            view.setText(currencyService.getCurrencyFormatted(currencyId, amount.toDouble()));
+            view.setText(currencyService.getCurrencyFormatted(currencyId, amount));
         }
 
         view.setTag(amount);
